Added face anonymization and crowd counting.
Made a single main script that runs both the scripts separately to minimize GPU memory usage. The initial image is anonymized and saved. Then the crowd counting script is ran and it takes that image, counts the people, saves the image in logs/ with dots around the detected heads and prints the number of detected people in the image.